requirements.txt
Django==1.8.6
djangorestframework==3.3.1
wheel==0.26.0
